+++ pkg/server/graph/detached.go
@@ -4,9 +4,10 @@ import (
 	"context"
 	"crypto/md5"
 	"encoding/hex"
-	"github.com/jjzcru/elk/pkg/server/graph/model"
 	"sync"
 	"time"
+
+	"github.com/jjzcru/elk/pkg/server/graph/model"
 )
 
 type detachedContext struct {
@@ -14,8 +15,13 @@ type detachedContext struct {
 	cancel context.CancelFunc
 }
 
+// ServerCtx stores the context on which the server is running
 var ServerCtx context.Context
+
+// DetachedTasksMap links detached task with an id
 var DetachedTasksMap = make(map[string]*model.DetachedTask)
+
+// DetachedCtxMap stores the context of each task using an id
 var DetachedCtxMap = make(map[string]*detachedContext)
 
 func getDetachedTaskID() string {
@@ -30,6 +36,14 @@ func getDetachedTaskID() string {
 	}
 
 	return id
+}
+
+func getResponseFromDetached(id string) *model.DetachedTask {
+	return DetachedTasksMap[id]
+}
+
+func updateDetachedTask(id string, task *model.DetachedTask) {
+	DetachedTasksMap[id] = task
 }
 
 // CancelDetachedTasks call cancel on all the context
@@ -61,14 +75,29 @@ func CancelDetachedTasks() {
 }
 
 func delayStart(delay *time.Duration, start *time.Time) {
+	sleepDuration := getDelayDuration(delay, start)
+	if sleepDuration > 0 {
+		time.Sleep(sleepDuration)
+	}
+}
+
+func getDelayDuration(delay *time.Duration, start *time.Time) time.Duration {
 	var startDuration time.Duration
 	var delayDuration time.Duration
-	var sleepDuration time.Duration
 
 	if start != nil {
 		now := time.Now()
-		startTime := *start
+		var startTime time.Time
 
+		if start.Before(now) {
+			startTime = time.Date(now.Year(), now.Month(), now.Day(),
+				start.Hour(), start.Minute(), start.Second(),
+				start.Nanosecond(), start.Location())
+
+			startTime.Add(24 * time.Hour)
+		} else {
+			startTime = *start
+		}
 		startDuration = startTime.Sub(now)
 	}
 
@@ -78,17 +107,14 @@ func delayStart(delay *time.Duration, start *time.Time) {
 
 	if startDuration > 0 && delayDuration > 0 {
 		if startDuration > delayDuration {
-			sleepDuration = startDuration
-		} else {
-			sleepDuration = delayDuration
+			return startDuration
 		}
+		return delayDuration
 	} else if startDuration > 0 {
-		sleepDuration = startDuration
+		return startDuration
 	} else if delayDuration > 0 {
-		sleepDuration = delayDuration
+		return delayDuration
 	}
 
-	if sleepDuration > 0 {
-		time.Sleep(sleepDuration)
-	}
+	return 0
 }